+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to
   - [#2147](https://github.com/iovisor/bpftrace/pull/2147)
 - Improve include paths resolution
   - [#2149](https://github.com/iovisor/bpftrace/pull/2149)
+- Add builtin function: `bswap`
+  - [#2166](https://github.com/iovisor/bpftrace/pull/2166)
 
 #### Changed
 #### Deprecated